+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- [all] update dependencies
+- [all] update lodash imports to allow rollup to treeshake better
+- [form] add `validate` function to form
 - [slider] _breaking_ change `next` behavior so that it calls the submit instead of going to the last slide when no viable candidates are found
 - [field] allow for `asField` to set a default value for a field type
 - [field] update file structures to appease the linter